fix(TBD-10885): refacto date and timestamp handling with dataset
